LeetCode problem 329 Longest Increasing Path
This is my solution to LeetCode problem 329, Longest Increasing Path in a Matrix. Given an m by n matrix of integers, the program returns the length of the longest path that can be taken through the matrix moving from entries of lower value to adjacent entries of greater value. Permissible moves are up, down, left, or right. Both m and n are between 1 and 200, and the entries are bounded by 0 and 2^32 - 1. 

The program treats the matrix as a directed acyclic graph, with a directed edge from entry A to entry B iff A and B are adjacent and A has a lower value than B, and topologically sorts the DAG. Then the length of the longest increasing path can be found in linear time.