LeetCode 6 Zigzag Conversion
This is my solution to LeetCode problem 6, Zigzag Conversion. Given a string s and an integer n, the program returns a string containing the same characters read in reading order from an arrangement of s into a horizontal zigzag of height n. The example s='ab', n=1 is worked out in comments.